% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{estep}
\alias{estep}
\title{E-step}
\usage{
estep(y, w, x.control, x.sensitive, x.outcome, x.misreport, treat, J,
  par.sensitive, par.control, par.outcome, par.outcome.aux, par.misreport, d,
  sensitive.response, model.misreport, o, trials, outcome.model, weight,
  outcome.constrained, control.constraint, respondentType, misreport.treatment)
}
\description{
The expectation step in the EM algorithm called by \code{\link{listExperiment}}.
}
\keyword{internal}

