% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{logAdd}
\alias{logAdd}
\title{Add two numbers}
\usage{
logAdd(lx, ly)
}
\arguments{
\item{x}{log of the first number.}

\item{y}{log of the second number.}
}
\value{
Log of the sum of \code{exp(x)} and \code{exp(y)}.
}
\description{
The function \code{logAdd} adds together two numbers using their log
to prevent under-/over-flow
}
\keyword{internal}

