% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{mstepControl}
\alias{mstepControl}
\title{Control-items sub-model m-step}
\usage{
mstepControl(y, treat, J, x.control, w, d, sensitive.response, weight,
  model.misreport, control.constraint)
}
\description{
The maximization step in the EM algorithm called by \code{\link{listExperiment}}
for the control-items sub-model.
}
\keyword{internal}

