% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{mstepMisreport}
\alias{mstepMisreport}
\title{Misreport sub-model m-step}
\usage{
mstepMisreport(y, x.misreport, w, treat, misreport.treatment, weight)
}
\description{
The maximization step in the EM algorithm called by \code{\link{listExperiment}}
for the misreport sub-model.
}
\keyword{internal}

