% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{mstepOutcome}
\alias{mstepOutcome}
\title{Outcome sub-model m-step}
\usage{
mstepOutcome(y, treat, x.outcome, w, d, sensitive.response, o, trials, weight,
  outcome.model, model.misreport, outcome.constrained, control.constraint)
}
\description{
The maximization step in the EM algorithm called by \code{\link{listExperiment}}
for the outcome sub-model.
}
\keyword{internal}

