% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{mstepSensitive}
\alias{mstepSensitive}
\title{Sensitive-item sub-model m-step}
\usage{
mstepSensitive(y, treat, x.sensitive, w, d, sensitive.response, weight,
  model.misreport)
}
\description{
The maximization step in the EM algorithm called by \code{\link{listExperiment}}
for the sensitive-item sub-model.
}
\keyword{internal}

