% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ListExperimentFunctions.R
\name{print.listExperiment}
\alias{print.listExperiment}
\title{Print object summary of listExperiment class}
\usage{
\method{print}{listExperiment}(x, ...)
}
\arguments{
\item{x}{Object of class "listExperiment".}

\item{...}{Additional arguments.}
}
\description{
Calls \code{\link{summary.listExperiment}}.
}
\details{
Prints the object summary of the listExperiment class by calling the
\code{\link{summary.listExperiment}} function.
}

